% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_methylation_raw}
\alias{data_methylation_raw}
\title{Example DNA methylation data}
\format{
A data frame with CpG sites in rows and samples in columns.
}
\usage{
data_methylation_raw
}
\description{
A  dataset containing raw beta methylation values for CpG sites.
Values range between 0 and 1, extracted from the public dataset GSE73103.
}
\keyword{datasets}
