% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functintermethsnp_continu.R
\name{fitMethSNPfunctInter_continu}
\alias{fitMethSNPfunctInter_continu}
\title{Functional Interaction Model between SNPs and Methylation for Continuous Outcomes}
\usage{
fitMethSNPfunctInter_continu(
  Y,
  W,
  G,
  mat_methyl,
  pos_snp,
  rho,
  nb_fct,
  kernel_name,
  region_range,
  RGN = seq(0, 1, length.out = 1000)
)
}
\arguments{
\item{Y}{Numeric vector: continuous response variable}

\item{W}{Matrix of covariates (individuals x covariates)}

\item{G}{Matrix of SNPs (individuals x SNPs, values in 0/1/2)}

\item{mat_methyl}{Matrix of methylation curves (individuals x grid points),
as obtained using the \code{estim_methyl_curve()} function.}

\item{pos_snp}{Real SNP positions}

\item{rho}{Numeric: kernel parameter controlling the weight decay}

\item{nb_fct}{Number of basis functions}

\item{kernel_name}{One of "convex", "gaussian", or "linear"}

\item{region_range}{Vector of two elements: start and end positions of CpGs in region of interest}

\item{RGN}{Numeric vector. Grid of values in [0,1] used for curve estimation
(default = seq(0,1,length.out = 1000)).}
}
\value{
A list with:
\itemize{
\item \code{pvalue}: Global p-value of the interaction test,
\item \code{teta}: Vector of estimated model coefficients,
\item \code{sig2}: Estimated residual variance,
\item \code{lambda}: Estimated smoothing parameter,
\item \code{object_gam}: Object of class \code{gam} (from \pkg{mgcv}).
Can be used with \code{summary()} to obtain individual tests,
and \code{plot()} to visualize the estimated curve \eqn{\hat{\delta}(t)}.
}
}
\description{
This function fits a functional regression model to assess the global interaction between SNP genotypes and DNA methylation profiles on continuous outcomes. A set of kernel-weighted interaction terms is constructed, and a functional smooth term is estimated using penalized splines. A Wald-type test is performed to evaluate the global significance of the interaction effects.
}
\examples{
\dontrun{
# Simulate data
set.seed(123)
n <- 100  # nombre d'individus
p <- 3    # nombre de SNPs
m <- 500  # nombre de points CpG

Y <- rnorm(n)
W <- matrix(rnorm(n * 2), ncol = 2)
colnames(W) <- c("age", "sex")

G <- matrix(sample(0:2, n * p, replace = TRUE), ncol = p)
colnames(G) <- c("rs101", "rs202", "rs303")

# Example of methylation curve estimation
mat_methyl <- matrix(runif(n * m), nrow = n)  # should be obtained beforehand.

pos_snp_bp <- c(120000, 150000, 180000)
region_range <- c(100000, 200000)

res <- fitMethSNPfunctInter_continu(
  Y = Y,
  W = W,
  G = G,
  mat_methyl = mat_methyl,
  pos_snp = pos_snp_bp,
  rho = 10,
  nb_fct = 10,
  kernel_name = "gaussian",
  region_range = region_range
)

res$pvalue  # Pvalue for the globalinteraction test
}
}
