% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functintermethsnp_generalized.R
\name{fitMethSNPfunctInter_generalized}
\alias{fitMethSNPfunctInter_generalized}
\title{Global Test for SNP–Methylation Interactions on Generalized Outcomes via Functional Regression}
\usage{
fitMethSNPfunctInter_generalized(
  Y,
  W,
  G,
  mat_methyl,
  pos_snp,
  rho,
  nb_fct,
  kernel_name,
  region_range,
  RGN = seq(0, 1, length.out = 1000),
  family = binomial()
)
}
\arguments{
\item{Y}{Response variable (vector), assumed from exponential family (e.g., binary, Poisson)}

\item{W}{Matrix of covariates (individuals x covariates)}

\item{G}{Matrix of SNPs (individuals x SNPs, values in 0/1/2)}

\item{mat_methyl}{Matrix of methylation curves (individuals x grid points)}

\item{pos_snp}{SNP base pair positions (numeric vector)}

\item{rho}{Numeric: kernel parameter controlling the weight decay}

\item{nb_fct}{Number of basis functions in the GAM}

\item{kernel_name}{One of "convex", "gaussian", or "linear"}

\item{region_range}{Vector of two elements: start and end positions in bp of CpGs in region of interest}

\item{RGN}{Numeric vector. Grid of values in [0,1] used for curve estimation
(default = seq(0,1,length.out = 1000)).}

\item{family}{A family object describing the error distribution (e.g., binomial(), poisson())}
}
\value{
A list with:
\itemize{
\item \code{pvalue}: p-value from the likelihood ratio test comparing models with and without interaction terms,
\item \code{teta}: vector of estimated coefficients from the full model,
\item \code{lambda}: Estimated smoothing parameter.,
\item \code{object_gam}: the fitted \code{mgcv::gam} object (full model).
}
}
\description{
This function fits a functional regression to assess the global interaction between SNP genotypes and DNA methylation profiles on non-Gaussian outcomes, such as binary or count responses.
A likelihood ratio test is performed to evaluate the significance of the interaction terms.
}
\examples{
\dontrun{
# Simuler des données
set.seed(123)
n <- 100  # nombre d'individus
p <- 3    # nombre de SNPs
m <- 500  # nombre de points CpG

Y <- rbinom(n, size = 1, prob = 0.4)
W <- matrix(rnorm(n * 2), ncol = 2)
colnames(W) <- c("age", "sex")

G <- matrix(sample(0:2, n * p, replace = TRUE), ncol = p)
colnames(G) <- c("rs101", "rs202", "rs303")

mat_methyl <- matrix(runif(n * m), nrow = n)

pos_snp_bp <- c(120000, 150000, 180000)
region_range <- c(100000, 200000)

res <- fitMethSNPfunctInter_generalized(
  Y = Y,
  W = W,
  G = G,
  mat_methyl = mat_methyl,
  pos_snp = pos_snp_bp,
  rho = 10,
  nb_fct = 10,
  kernel_name = "gaussian",
  region_range = region_range,
  family=binomial()
)

res$pvalue  #pvalue for the global interaction testing
}
}
