% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_covar}
\alias{mat_covar}
\title{Example covariate matrix}
\format{
A matrix with individuals in rows and covariates in columns.
}
\usage{
mat_covar
}
\description{
Contains potential confounders such as age and sex.
}
\keyword{datasets}
