% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_methyl}
\alias{mat_methyl}
\title{Example methylation curves}
\format{
A matrix with individuals in rows and grid points in columns.
}
\usage{
mat_methyl
}
\description{
Estimated methylation curves for individuals over a genomic region.
}
\keyword{datasets}
