% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_snp}
\alias{mat_snp}
\title{Example SNP genotype matrix}
\format{
A matrix with individuals in rows and SNPs in columns.
}
\usage{
mat_snp
}
\description{
SNP genotypes coded as 0, 1, 2 for individuals.
}
\keyword{datasets}
