% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos_snp}
\alias{pos_snp}
\title{Example SNP base pair positions}
\format{
A numeric vector with length equal to the number of SNPs.
}
\usage{
pos_snp
}
\description{
Physical base-pair positions of the SNPs.
}
\keyword{datasets}
