% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resp_binaire}
\alias{resp_binaire}
\title{Binary response (0/1)}
\format{
A numeric vector of length equal to the number of individuals, with values 0 or 1.
}
\usage{
resp_binaire
}
\description{
A  binary outcome variable (e.g., obesity status).
}
\keyword{datasets}
