% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resp_continu}
\alias{resp_continu}
\title{Simulated continuous response}
\format{
A numeric vector of length equal to the number of individuals.
}
\usage{
resp_continu
}
\description{
A  continuous outcome variable.
}
\keyword{datasets}
