% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resp_pois}
\alias{resp_pois}
\title{Simulated Poisson response (counts)}
\format{
A numeric vector of length equal to the number of individuals.
}
\usage{
resp_pois
}
\description{
A  count outcome variable generated from a Poisson model.
}
\keyword{datasets}
